+++ domains/certificates/Query.API/API/API.csproj
@@ -1,4 +1,4 @@
-﻿<Project Sdk="Microsoft.NET.Sdk.Web">
+<Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
     <TargetFramework>net6.0</TargetFramework>
@@ -34,6 +34,11 @@
     <None Update="documentation.xml">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </None>
+  </ItemGroup>
+
+  <ItemGroup>
+    <Folder Include="Query.API\ApiModels\Requests\" />
+    <Folder Include="Query.API\ApiModels\Responses\" />
   </ItemGroup>
 
 </Project>
