+++ domains/certificates/Query.API/API/Query.API/ApiModels/Responses/Certificate.cs
@@ -1,7 +1,7 @@
 using System;
 using System.Text.Json.Serialization;
 
-namespace API.Query.API.ApiModels;
+namespace API.Query.API.ApiModels.Responses;
 
 public class Certificate
 {
