+++ domains/certificates/Query.API/API/Query.API/Controllers/CertificatesController.cs
@@ -1,6 +1,6 @@
 using System.Security.Claims;
 using System.Threading.Tasks;
-using API.Query.API.ApiModels;
+using API.Query.API.ApiModels.Responses;
 using API.Query.API.Projections;
 using Marten;
 using Microsoft.AspNetCore.Authorization;
